requirements.txt
amqp==5.1.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
idna==3.4
kombu==5.2.4
libretranslatepy==2.1.1
lxml==4.9.2
mysql-connector-python==8.0.33
prompt-toolkit==3.0.38
protobuf==3.20.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4
translate==3.6.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
